mssql backend progress

+++ include/database/mssql/mssql_types.hpp
@@ -30,21 +30,21 @@ namespace mssql {
 template <class T> 
 struct type_traits; 
 
-template <> struct type_traits<char> { inline static const char* type_string() { return "INTEGER"; } }; 
-template <> struct type_traits<short> { inline static const char* type_string() { return "INTEGER"; } };
-template <> struct type_traits<int> { inline static const char* type_string() { return "INTEGER"; } };
-template <> struct type_traits<long> { inline static const char* type_string() { return "INTEGER"; } };
-template <> struct type_traits<unsigned char> { inline static const char* type_string() { return "INTEGER"; } };
-template <> struct type_traits<unsigned short> { inline static const char* type_string() { return "INTEGER"; } };
-template <> struct type_traits<unsigned int> { inline static const char* type_string() { return "INTEGER"; } };
-template <> struct type_traits<unsigned long> { inline static const char* type_string() { return "INTEGER"; } };
-template <> struct type_traits<bool> { inline static const char* type_string() { return "INTEGER"; } };
-template <> struct type_traits<float> { inline static const char* type_string() { return "REAL"; } };
-template <> struct type_traits<double> { inline static const char* type_string() { return "REAL"; } };
+template <> struct type_traits<char> { inline static const char* type_string() { return "TINYINT"; } }; 
+template <> struct type_traits<short> { inline static const char* type_string() { return "SMALLINT"; } };
+template <> struct type_traits<int> { inline static const char* type_string() { return "INT"; } };
+template <> struct type_traits<long> { inline static const char* type_string() { return "INT"; } };
+template <> struct type_traits<unsigned char> { inline static const char* type_string() { return "TINYINT"; } };
+template <> struct type_traits<unsigned short> { inline static const char* type_string() { return "SMALLINT"; } };
+template <> struct type_traits<unsigned int> { inline static const char* type_string() { return "INT"; } };
+template <> struct type_traits<unsigned long> { inline static const char* type_string() { return "INT"; } };
+template <> struct type_traits<bool> { inline static const char* type_string() { return "TINYINT"; } };
+template <> struct type_traits<float> { inline static const char* type_string() { return "FLOAT"; } };
+template <> struct type_traits<double> { inline static const char* type_string() { return "FLOAT"; } };
 template <> struct type_traits<varchar_base> { inline static const char* type_string() { return "VARCHAR"; } };
 template <> struct type_traits<const char*> { inline static const char* type_string() { return "VARCHAR"; } };
 template <> struct type_traits<std::string> { inline static const char* type_string() { return "TEXT"; } };
-template <> struct type_traits<object_base_ptr> { inline static const char* type_string() { return "INTEGER"; } };
+template <> struct type_traits<object_base_ptr> { inline static const char* type_string() { return "INT"; } };
 
 class mssql_types
 {
